Only link executable on OSX
really fixes #2

+++ sublime3/install.sh
@@ -3,9 +3,11 @@
 
 . functions.sh
 
-if [ `uname -s` == 'Darwin' ]; then
+UNAME=$(uname -s)
+
+if [ "${UNAME}" == "Darwin" ]; then
     sublime_dir=~/Library/Application\ Support/Sublime\ Text\ 3
-elif [ `uname -s` == "Linux" ]; then
+elif [ "${UNAME}" == "Linux" ]; then
     sublime_dir=~/.config/sublime-text-3
 else
     fail "unsupported operating system"
@@ -14,7 +16,7 @@ fi
 if [ ! -L "$sublime_dir/Packages/User" ]; then
     mv "$sublime_dir/Packages/User" "$sublime_dir/Packages/User.backup"
     ln -s ~/.dotfiles/sublime3/User "$sublime_dir/Packages/User"
-    if [ ! -L /usr/local/bin/subl ]; then
+    if [ "${UNAME}" == "Darwin" ] && [ ! -L /usr/local/bin/subl ]; then
         ln -s /Applications/Sublime\ Text.app/Contents/SharedSupport/bin/subl /usr/local/bin/
     fi
     success "sublime3 installed"
